update to Chromium 146.0.7680.31

+++ chrome/browser/paint_preview/services/paint_preview_tab_service.cc
@@ -219,18 +219,17 @@ void PaintPreviewTabService::AuditArtifacts(
 #if BUILDFLAG(IS_ANDROID)
 void PaintPreviewTabService::CaptureTabAndroid(
     JNIEnv* env,
-    jint j_tab_id,
+    int32_t j_tab_id,
     const base::android::JavaRef<jobject>& j_web_contents,
     bool j_accessibility_enabled,
-    jfloat j_page_scale_factor,
-    jint j_x,
-    jint j_y,
+    float j_page_scale_factor,
+    int32_t j_x,
+    int32_t j_y,
     const base::android::JavaRef<jobject>& j_callback) {
   content::WebContents* web_contents =
       content::WebContents::FromJavaWebContents(j_web_contents);
   CaptureTab(static_cast<int>(j_tab_id), web_contents, j_accessibility_enabled,
-             static_cast<float>(j_page_scale_factor), static_cast<int>(j_x),
-             static_cast<int>(j_y),
+             j_page_scale_factor, static_cast<int>(j_x), static_cast<int>(j_y),
              base::BindOnce(
                  &JavaBooleanCallbackAdapter,
                  base::BindOnce(
@@ -238,12 +237,12 @@ void PaintPreviewTabService::CaptureTabAndroid(
                      base::android::ScopedJavaGlobalRef<jobject>(j_callback))));
 }
 
-void PaintPreviewTabService::TabClosedAndroid(JNIEnv* env, jint j_tab_id) {
+void PaintPreviewTabService::TabClosedAndroid(JNIEnv* env, int32_t j_tab_id) {
   TabClosed(static_cast<int>(j_tab_id));
 }
 
 bool PaintPreviewTabService::HasCaptureForTabAndroid(JNIEnv* env,
-                                                     jint j_tab_id) {
+                                                     int32_t j_tab_id) {
   return static_cast<bool>(HasCaptureForTab(static_cast<int>(j_tab_id)));
 }
